celery: spostato codice relativo alla posta nell'app posta
